Use this as current recording system
Added try except blocks to a lot of the functions that perform tasks such as video acquiring, posting to APIs, connecting to database, connecting to MQTT broker. 

The new system records videos for 10 seconds. This system is able to continuously record videos and trim videos that exceed 10 seconds and delete their original videos.